FIX: compiler warnings
git-svn-id: https://svn.code.sf.net/p/flashfolder/code/trunk@249 a429ba1a-122f-0410-b780-ea7094eed80e

+++ FFLib/MsoFileDlgHook.cpp
@@ -196,7 +196,7 @@ bool MsoFileDlgHook::EnterFilenameEditText( LPCTSTR pText )
 
 	//--- send the prepared input stream ---
 
-	::SendInput( vinp.size(), &vinp[0], sizeof(INPUT) );
+	::SendInput( (UINT) vinp.size(), &vinp[0], sizeof(INPUT) );
 
 	return true;
 }
